services/watchguard/requirements.txt
# WatchGuard MVP Dependencies

# HTTP client
httpx>=0.27.0

# HTML parsing and text extraction
beautifulsoup4>=4.12.0
lxml>=5.0.0
readability-lxml>=0.8.1

# CLI
click>=8.1.0

# Configuration
pyyaml>=6.0.0

# Logging
structlog>=24.0.0

# Retry logic
tenacity>=8.2.0

# Phase 2: Playwright (uncomment when needed)
# playwright>=1.40.0

# Phase 3: PDF extraction (uncomment when needed)
# pdfplumber>=0.10.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0